Refactor: uddapte folder structure
