+++ TodoItems.Core/Constant.cs
@@ -1,0 +1,8 @@
+﻿namespace TodoItems.Core
+{
+    internal static class Constant
+    {
+        public const int MAX_TODOITEMS_PER_DUE_DATE = 8;
+        public const int MAX_MODIFICATION_PER_TODOITEM_PER_DAY = 3;
+    }
+}
