--- TodoItems.Core/User.cs
@@ -1,11 +1,0 @@
-﻿namespace TodoItems.Core;
-
-public class User
-{
-    public List<TodoItem> TodoItems { get; set; }
-    public User()
-    {
-        TodoItems = new List<TodoItem>();
-    }
-}
-
